+++ CHANGES.md
@@ -14,6 +14,7 @@ Fixed Issues:
 * [#12243](http://dev.ckeditor.com/ticket/12243): Fixed: Text formatting lost while pasting from Word. Thanks to [Alin Purcaru](https://github.com/mesmerizero)!
 * [#12273](http://dev.ckeditor.com/ticket/12273): Fixed: Applying block style in description list breaks it.
 * [#12281](http://dev.ckeditor.com/ticket/12281): Fixed: Minor syntax issue in CSS files.
+* [#12178](http://dev.ckeditor.com/ticket/12178): [Blink/Webkit] Fixed: Iterator does not return block if selection is located at the end of it.
 
 ## CKEditor 4.4.3